Move ASNField from dcim to ipam
