+++ netbox/dcim/fields.py
@@ -1,10 +1,8 @@
 from django.contrib.postgres.fields import ArrayField
 from django.core.exceptions import ValidationError
-from django.core.validators import MinValueValidator, MaxValueValidator
 from django.db import models
 from netaddr import AddrFormatError, EUI, eui64_unix_expanded, mac_unix_expanded
 
-from ipam.constants import BGP_ASN_MAX, BGP_ASN_MIN
 from .lookups import PathContains
 
 __all__ = (
@@ -26,22 +24,6 @@ class eui64_unix_expanded_uppercase(eui64_unix_expanded):
 #
 # Fields
 #
-
-class ASNField(models.BigIntegerField):
-    description = "32-bit ASN field"
-    default_validators = [
-        MinValueValidator(BGP_ASN_MIN),
-        MaxValueValidator(BGP_ASN_MAX),
-    ]
-
-    def formfield(self, **kwargs):
-        defaults = {
-            'min_value': BGP_ASN_MIN,
-            'max_value': BGP_ASN_MAX,
-        }
-        defaults.update(**kwargs)
-        return super().formfield(**defaults)
-
 
 class MACAddressField(models.Field):
     description = "PostgreSQL MAC Address field"
