+++ netbox/ipam/constants.py
@@ -2,10 +2,6 @@ from django.db.models import Q
 
 from .choices import FHRPGroupProtocolChoices, IPAddressRoleChoices
 
-# BGP ASN bounds
-BGP_ASN_MIN = 1
-BGP_ASN_MAX = 2**32 - 1
-
 
 #
 # VRFs
