+++ netbox/ipam/fields.py
@@ -1,9 +1,20 @@
 from django.core.exceptions import ValidationError
+from django.core.validators import MinValueValidator, MaxValueValidator
 from django.db import models
 from netaddr import AddrFormatError, IPNetwork
 
 from . import lookups, validators
 from .formfields import IPNetworkFormField
+
+__all__ = (
+    'ASNField',
+    'IPAddressField',
+    'IPNetworkField',
+)
+
+# BGP ASN bounds
+BGP_ASN_MIN = 1
+BGP_ASN_MAX = 2**32 - 1
 
 
 class BaseIPField(models.Field):
@@ -93,3 +104,19 @@ IPAddressField.register_lookup(lookups.NetIn)
 IPAddressField.register_lookup(lookups.NetHostContained)
 IPAddressField.register_lookup(lookups.NetFamily)
 IPAddressField.register_lookup(lookups.NetMaskLength)
+
+
+class ASNField(models.BigIntegerField):
+    description = "32-bit ASN field"
+    default_validators = [
+        MinValueValidator(BGP_ASN_MIN),
+        MaxValueValidator(BGP_ASN_MAX),
+    ]
+
+    def formfield(self, **kwargs):
+        defaults = {
+            'min_value': BGP_ASN_MIN,
+            'max_value': BGP_ASN_MAX,
+        }
+        defaults.update(**kwargs)
+        return super().formfield(**defaults)
